Factor out SimpleChannelInboundHandlerInstrumented
